requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.115.12
finance-datareader==0.9.96
h11==0.16.0
idna==3.10
lxml==5.4.0
narwhals==1.41.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
plotly==6.1.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
requests-file==2.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
